% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{add-vh-lines}
\alias{add-vh-lines}
\alias{add_hline}
\alias{add_vline}
\title{Add horizontal or vertical line}
\usage{
add_hline(ax, value, color = "#000", dash = 0, label = NULL, ...)

add_vline(ax, value, color = "#000", dash = 0, label = NULL, ...)
}
\arguments{
\item{ax}{An \code{\link[=apexchart]{apexchart()}} \code{htmlwidget} object.}

\item{value}{Vector of position for the line(s).}

\item{color}{Color(s) of the line(s).}

\item{dash}{Creates dashes in borders of SVG path.
A higher number creates more space between dashes in the border.
Use \code{0} for plain line.}

\item{label}{Add a label to the shade, use a \code{character}
or see \code{\link{label}} for more controls.}

\item{...}{Additional arguments, see
\url{https://apexcharts.com/docs/options/annotations/} for possible options.}
}
\value{
An \code{\link[=apexchart]{apexchart()}} \code{htmlwidget} object.
}
\description{
Add horizontal or vertical line
}
\examples{

library(apexcharter)

# On a column chart
unhcr_ts \%>\%
  subset(year == 2017 & population_type == "Asylum-seekers") \%>\%
  apex(
    aes(continent_origin, n),
    "column"
  ) \%>\%
  add_hline(value = 5e5)


# On a scatter chart
apex(
  data = cars,
  aes(speed, dist),
  "scatter"
) \%>\%
  add_hline(value = mean(cars$dist)) \%>\%
  add_vline(value = mean(cars$speed))

# With labels
apex(
  data = cars,
  aes(speed, dist),
  "scatter"
) \%>\%
  add_hline(
    value = mean(cars$dist),
    label = "Mean of dist"
  ) \%>\%
  add_vline(
    value = mean(cars$speed),
    label = label(
      text = "Mean of speed",
      borderColor = "red"
    )
  )
}
