% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex.R
\name{apex}
\alias{apex}
\title{Quick Apex Chart}
\usage{
apex(data, mapping, type = "column", ..., auto_update = TRUE,
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Default dataset to use for chart. If not already a \code{data.frame}, it will be coerced to with \code{as.data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart}

\item{type}{Specify the chart type. Available Options: \code{"column"}, \code{"bar"}, \code{"line"},
\code{"area"}, \code{"spline"}, \code{"pie"}, \code{"donut"}, \code{"radialBar"}, \code{"radar"}, \code{"scatter"}, \code{"heatmap"}.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{auto_update}{In Shiny application, update existing chart rather than generating new one.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\description{
Quick Apex Chart
}
