% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apexcharter-package.R
\name{apexcharter-exports}
\alias{apexcharter-exports}
\alias{\%>\%}
\alias{aes}
\alias{vars}
\alias{label_value}
\alias{JS}
\title{apexcharter exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the apexcharter package to avoid listing the magrittr
as Depends of apexcharter
}
