% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apexcharter.R
\name{apexcharter}
\alias{apexcharter}
\title{Create a apexcharts.js widget}
\usage{
apexcharter(data = NULL, ax_opts = list(), width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Default dataset to use for chart. If not already a \code{data.frame}, it will be coerced to with \code{as.data.frame}.}

\item{ax_opts}{A \code{list} in JSON format with chart parameters.#'}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Create a apexcharts.js widget
}
\examples{

library(apexcharter)

apexcharter(ax_opts = list(
  chart = list(type = "bar"),
  series = list(list(
    name = "Example",
    data = sample(1:100, 5)
  )),
  xaxis = list(categories = LETTERS[1:5])
))
}
