% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{apexchartProxy}
\alias{apexchartProxy}
\title{Proxy for \code{apexchart}}
\usage{
apexchartProxy(shinyId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shinyId}{single-element character vector indicating the output ID of the
chart to modify (if invoked from a Shiny module, the namespace will be added
automatically)}

\item{session}{the Shiny session object to which the chart belongs; usually the
default value will suffice}
}
\description{
Allow to update a chart in Shiny application.
}
