% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax-series}
\alias{ax-series}
\alias{ax_series}
\alias{ax_series2}
\title{Add data to a chart}
\usage{
ax_series(ax, ...)

ax_series2(ax, l)
}
\arguments{
\item{ax}{An \code{\link[=apexchart]{apexchart()}} \code{htmlwidget} object.}

\item{...}{Lists containing data to plot, typically list with two items: \code{name} and \code{data}.}

\item{l}{A list.}
}
\value{
An \code{\link[=apexchart]{apexchart()}} \code{htmlwidget} object.
}
\description{
Add data to a chart
}
\examples{

# One serie
apexchart() \%>\% 
  ax_series(list(
    name = "rnorm",
    data = rnorm(10)
  ))

# Two series
apexchart() \%>\% 
  ax_series(
    list(
      name = "rnorm 1",
      data = rnorm(10)
    ),
    list(
      name = "rnorm 2",
      data = rnorm(10)
    )
  )
}
