% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_annotations}
\alias{ax_annotations}
\title{Annotations properties}
\usage{
ax_annotations(ax, position = NULL, yaxis = NULL, xaxis = NULL,
  points = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{position}{}

\item{yaxis}{}

\item{xaxis}{}

\item{points}{}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Annotations properties
}
