% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_chart}
\alias{ax_chart}
\title{Chart parameters}
\usage{
ax_chart(ax, type = NULL, stacked = NULL, stackType = NULL,
  animations = NULL, background = NULL, foreColor = NULL,
  dropShadow = NULL, events = NULL, offsetX = NULL, offsetY = NULL,
  scroller = NULL, selection = NULL, sparkline = NULL,
  toolbar = NULL, zoom = NULL, width = NULL, height = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{type}{Specify the chart type. Available Options: \code{"bar"}, \code{"column"}, \code{"line"},
\code{"histogram"}, \code{"pie"}, \code{"donut"}, \code{"radialBar"}, \code{"scatter"}, \code{"bubble"}, \code{"heatmap"}.}

\item{stacked}{Logical. Enables stacked option for axis charts.}

\item{stackType}{When stacked, should the stacking be percentage based or normal stacking. Available options: \code{"normal"} or \code{"100\%"}}

\item{animations}{List.}

\item{background}{Background color for the chart area. If you want to set background with css, use \code{.apexcharts-canvas} to set it.}

\item{foreColor}{Sets the text color for the chart. Defaults to \code{#373d3f}.}

\item{dropShadow}{List.}

\item{events}{List.}

\item{offsetX}{}

\item{offsetY}{}

\item{scroller}{List.}

\item{selection}{List.}

\item{sparkline}{List. Sparkline hides all the elements of the charts other than the primary paths. Helps to visualize data in small areas. .}

\item{toolbar}{List.}

\item{zoom}{List.}

\item{width}{Width of the chart.}

\item{height}{Height of the chart.}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Chart parameters
}
