% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_dataLabels}
\alias{ax_dataLabels}
\title{Labels on data}
\usage{
ax_dataLabels(ax, enabled = NULL, textAnchor = NULL, offsetX = NULL,
  offsetY = NULL, style = NULL, dropShadow = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{enabled}{}

\item{textAnchor}{}

\item{offsetX}{}

\item{offsetY}{}

\item{style}{}

\item{dropShadow}{}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Labels on data
}
