% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{ax_facet_wrap}
\alias{ax_facet_wrap}
\title{Facet wrap for ApexCharts}
\usage{
ax_facet_wrap(ax, vars, nrow = NULL, ncol = NULL, chart_height = "300px")
}
\arguments{
\item{ax}{An \code{apexcharts} \code{htmlwidget} object.}

\item{vars}{Variable(s) to use for facetting, wrapped in \code{vars(...)}.}

\item{nrow, ncol}{Number of row and column in output matrix.}

\item{chart_height}{Individual chart height.}
}
\value{
An \code{apexcharts} \code{htmlwidget} object.
}
\description{
Facet wrap for ApexCharts
}
