% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_fill}
\alias{ax_fill}
\title{Fill property}
\usage{
ax_fill(ax, type = NULL, colors = NULL, opacity = NULL,
  gradient = NULL, image = NULL, pattern = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{type}{Whether to fill the paths with solid colors or gradient.
Available options: \code{"solid"}, \code{"gradient"}, \code{"pattern"} or \code{"image"}.}

\item{colors}{Colors to fill the svg paths..}

\item{opacity}{Opacity of the fill attribute.}

\item{gradient}{A list of parameters.}

\item{image}{A list of parameters.}

\item{pattern}{A list of parameters.}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Fill property
}
\note{
See \url{https://apexcharts.com/docs/options/fill/}
}
