% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_labels}
\alias{ax_labels}
\title{Alternative axis labels}
\usage{
ax_labels(ax, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{...}{}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Alternative axis labels
}
