% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_legend}
\alias{ax_legend}
\title{Legend properties}
\usage{
ax_legend(ax, show = NULL, floating = NULL, position = NULL,
  horizontFalAlign = NULL, verticalAlign = NULL, fontSize = NULL,
  textAnchor = NULL, offsetY = NULL, offsetX = NULL, formatter = NULL,
  labels = NULL, markers = NULL, itemMargin = NULL,
  containerMargin = NULL, onItemClick = NULL, onItemHover = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{show}{}

\item{floating}{}

\item{position}{}

\item{verticalAlign}{}

\item{fontSize}{}

\item{textAnchor}{}

\item{offsetY}{}

\item{offsetX}{}

\item{formatter}{}

\item{labels}{}

\item{markers}{}

\item{itemMargin}{}

\item{containerMargin}{}

\item{onItemClick}{}

\item{onItemHover}{}

\item{...}{Additional parameters.}

\item{horizontalAlign}{}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Legend properties
}
