% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_markers}
\alias{ax_markers}
\title{Markers properties}
\usage{
ax_markers(ax, discrete = NULL, size = NULL, colors = NULL,
  strokeColor = NULL, strokeWidth = NULL, strokeOpacity = NULL,
  fillOpacity = NULL, shape = NULL, radius = NULL, offsetX = NULL,
  offsetY = NULL, hover = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{discrete}{}

\item{size}{}

\item{colors}{}

\item{strokeColor}{}

\item{strokeWidth}{}

\item{strokeOpacity}{}

\item{fillOpacity}{}

\item{shape}{}

\item{radius}{}

\item{offsetX}{}

\item{offsetY}{}

\item{hover}{}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Markers properties
}
