% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_noData}
\alias{ax_noData}
\title{No data specification}
\usage{
ax_noData(ax, text = NULL, align = NULL, verticalAlign = NULL,
  offsetX = NULL, offsetY = NULL, style = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{text}{}

\item{align}{}

\item{verticalAlign}{}

\item{offsetX}{}

\item{offsetY}{}

\item{style}{}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
No data specification
}
