% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{ax_proxy_series}
\alias{ax_proxy_series}
\title{Proxy for updating series.}
\usage{
ax_proxy_series(proxy, newSeries, animate = TRUE)
}
\arguments{
\item{proxy}{A \code{apexchartProxy} \code{htmlwidget} object.}

\item{newSeries}{The series array to override the existing one.}

\item{animate}{Should the chart animate on re-rendering.}
}
\description{
Allows you to update the series array overriding the existing one.
}
\examples{
\dontrun{

if (interactive()) {



}

}
}
