% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_responsive}
\alias{ax_responsive}
\title{Responsive options}
\usage{
ax_responsive(ax, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{...}{}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Responsive options
}
