% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_states}
\alias{ax_states}
\title{Charts' states}
\usage{
ax_states(ax, normal = NULL, hover = NULL, active = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{normal}{List.}

\item{hover}{List.}

\item{active}{List.}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Charts' states
}
