% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_title}
\alias{ax_title}
\title{Chart's title}
\usage{
ax_title(ax, text = NULL, align = NULL, margin = NULL, offsetX = NULL,
  offsetY = NULL, floating = NULL, style = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{text}{}

\item{align}{}

\item{margin}{}

\item{offsetX}{}

\item{offsetY}{}

\item{floating}{}

\item{style}{}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Chart's title
}
