% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_xaxis}
\alias{ax_xaxis}
\title{X-axis options}
\usage{
ax_xaxis(ax, type = NULL, categories = NULL, labels = NULL,
  axisBorder = NULL, axisTicks = NULL, tickAmount = NULL, min = NULL,
  max = NULL, range = NULL, floating = NULL, position = NULL,
  title = NULL, crosshairs = NULL, tooltip = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{type}{Character. Available Options : \code{"categories"} and \code{"datetime"}.}

\item{categories}{Categories are labels which are displayed on the x-axis.}

\item{labels}{List.}

\item{axisBorder}{List.}

\item{axisTicks}{List.}

\item{tickAmount}{Number of Tick Intervals to show.}

\item{min}{Lowest number to be set for the x-axis. The graph drawing beyond this number will be clipped off.}

\item{max}{Highest number to be set for the x-axis. The graph drawing beyond this number will be clipped off.}

\item{range}{Range takes the max value of x-axis, subtracts the provided range value and gets the min value based on that.
So, technically it helps to keep the same range when min and max values gets updated dynamically.}

\item{floating}{Logical. Floating takes x-axis is taken out of normal flow and places x-axis on svg element directly,
similar to an absolutely positioned element. Set the offsetX and offsetY then to adjust the position manually}

\item{position}{Setting this option allows you to change the x-axis position. Available options: \code{"top"} and \code{"bottom"}.}

\item{title}{List.}

\item{crosshairs}{List.}

\item{tooltip}{List.}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
X-axis options
}
