% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blockbusters}
\alias{blockbusters}
\title{Worldwide Blockbusters 2019-1977}
\format{
A data frame with 183 observations and 3 variables.
}
\source{
Kaggle (\url{https://www.kaggle.com/datasets/narmelan/top-ten-blockbusters-20191977})
}
\usage{
blockbusters
}
\description{
The dataset contains data about about blockbusters gross (in millions USD) by genre and per year between 1977 and 2019.
}
\keyword{datasets}
