% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candles}
\alias{candles}
\title{Candlestick demo data}
\format{
A data frame with 60 observations and the following 5 variables:
\describe{
\item{\code{datetime}}{Timestamp.}
\item{\code{open}}{Open value.}
\item{\code{high}}{Highest value.}
\item{\code{low}}{Lowest value.}
\item{\code{close}}{Close value.}
}
}
\source{
Apexcharts (\url{https://apexcharts.com/javascript-chart-demos/candlestick-charts/basic/})
}
\usage{
candles
}
\description{
Candlestick demo data
}
\keyword{datasets}
