% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_paris}
\alias{climate_paris}
\title{Paris Climate}
\format{
A data frame with 12 observations and the following 3 variables:
\describe{
\item{\code{month}}{Month}
\item{\code{temperature}}{Temperature (in degree celsius).}
\item{\code{precipitation}}{Precipitation (in mm).}
}
}
\source{
Wikipedia (\url{https://fr.wikipedia.org/wiki/Climat_de_Paris})
}
\usage{
climate_paris
}
\description{
Average temperature and precipitation in Paris for the period 1971-2000.
}
\keyword{datasets}
