% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apexcharter.R
\name{config_update}
\alias{config_update}
\title{Configuration for auto update}
\usage{
config_update(
  series_animate = TRUE,
  update_options = FALSE,
  options_animate = TRUE,
  options_redrawPaths = TRUE,
  update_synced_charts = FALSE
)
}
\arguments{
\item{series_animate}{Should the chart animate on re-rendering.}

\item{update_options}{Update or not global options for chart.}

\item{options_animate}{Should the chart animate on re-rendering.}

\item{options_redrawPaths}{When the chart is re-rendered,
should it draw from the existing paths or completely redraw
the chart paths from the beginning. By default, the chart
is re-rendered from the existing paths.}

\item{update_synced_charts}{All the charts in a group should
also update when one chart in a group is updated.}
}
\description{
Configuration for auto update
}
