% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consumption}
\alias{consumption}
\title{Electricity consumption and forecasting}
\format{
A data frame with 120 observations and the following 3 variables:
\describe{
\item{\code{date}}{date.}
\item{\code{type}}{Type of data : realised or forecast.}
\item{\code{value}}{Value in giga-watt per hour.}
}
}
\source{
Rte (Electricity Transmission Network in France) (\url{https://data.rte-france.com/})
}
\usage{
consumption
}
\description{
Electricity consumption per day in France for january and february of year 2020.
}
\keyword{datasets}
