% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eco2mix}
\alias{eco2mix}
\title{eco2mix data}
\format{
A data frame with 3,033 observations and 3 variables.
}
\source{
Rte (Réseau et transport d'électricité) (\url{https://www.rte-france.com/donnees-publications/eco2mix-donnees-temps-reel} and \url{https://opendata.reseaux-energies.fr/})
}
\usage{
eco2mix
}
\description{
The dataset contains data about electricity consumption and production in France between 2012 and 2022.
}
\keyword{datasets}
