% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-options.R
\name{eventsOpts}
\alias{eventsOpts}
\title{Events options}
\usage{
eventsOpts(beforeMount = NULL, mounted = NULL, updated = NULL,
  clicked = NULL, selection = NULL, dataPointSelection = NULL,
  zoomed = NULL, scrolled = NULL, ...)
}
\arguments{
\item{beforeMount}{JS function. Fires before the chart has been drawn on screen.}

\item{mounted}{JS function. Fires after the chart has been drawn on screen.}

\item{updated}{JS function. Fires when the chart has been dynamically updated either with updateOptions() or updateSeries() functions.}

\item{clicked}{JS function. Fires when user clicks on any area of the chart.}

\item{selection}{JS function. Fires when user selects rect using the selection tool.}

\item{dataPointSelection}{JS function. Fires when user clicks on a datapoint.}

\item{zoomed}{JS function. Fires when user zooms in/out the chart using either the selection zooming tool or zoom in/out buttons.}

\item{scrolled}{JS function. Fires when user scrolls using either the pan tool or scroller.}

\item{...}{Additional parameters.}
}
\description{
Events options
}
\note{
See \url{https://apexcharts.com/docs/options/chart/events/}.
}
