% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-options.R
\name{heatmapOpts}
\alias{heatmapOpts}
\title{Heatmap options}
\usage{
heatmapOpts(radius = NULL, enableShades = NULL, shadeIntensity = NULL,
  colorScale = NULL, ...)
}
\arguments{
\item{radius}{Numeric. Radius of the rectangle inside heatmap.}

\item{enableShades}{Logical. Enable different shades of color depending on the value}

\item{shadeIntensity}{Numeric \code{[0,1]}. The intensity of the shades generated for each value.}

\item{colorScale}{List.}

\item{...}{Additional parameters.}
}
\description{
Use these options in \code{\link{ax_plotOptions}}.
}
\note{
See \url{https://apexcharts.com/docs/options/plotoptions/heatmap/}.
}
