% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{label}
\alias{label}
\title{Label for annotations}
\usage{
label(
  text = NULL,
  borderColor = NULL,
  borderWidth = NULL,
  textAnchor = NULL,
  position = NULL,
  offsetX = NULL,
  offsetY = NULL,
  background = NULL,
  color = NULL,
  fontSize = NULL,
  fontWeight = NULL,
  fontFamily = NULL,
  cssClass = NULL,
  padding = c(2, 5, 2, 5)
)
}
\arguments{
\item{text}{Text for the annotation label.}

\item{borderColor}{Border color for the label.}

\item{borderWidth}{Border width for the label.}

\item{textAnchor}{The alignment of text relative to label's drawing position.}

\item{position}{Available options: left or right.}

\item{offsetX}{Sets the left offset for annotation label.}

\item{offsetY}{Sets the top offset for annotation label.}

\item{background}{Background Color for the annotation label.}

\item{color}{ForeColor for the annotation label.}

\item{fontSize}{FontSize for the annotation label.}

\item{fontWeight}{Font-weight for the annotation label.}

\item{fontFamily}{Font-family for the annotation label.}

\item{cssClass}{A custom Css Class to give to the annotation label elements.}

\item{padding}{Padding for the label: top, right, bottom, left.}
}
\value{
A \code{list} that can be used in \code{\link{add_shade}}, \code{\link{add_point}},
\code{\link{add_event}}, \code{\link{add_event_marker}}.
}
\description{
Label for annotations
}
