% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-data.R
\name{parse_df}
\alias{parse_df}
\title{Convert a \code{data.frame} to a \code{list}}
\usage{
parse_df(data)
}
\arguments{
\item{data}{A \code{data.frame} or an object coercible to \code{data.frame}.}
}
\description{
Convert data to a format suitable for ApexCharts.js
}
