% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-options.R
\name{radialBar_opts}
\alias{radialBar_opts}
\title{Radial bar options}
\usage{
radialBar_opts(size = NULL, inverseOrder = NULL, startAngle = NULL,
  endAngle = NULL, offsetX = NULL, offsetY = NULL, hollow = NULL,
  track = NULL, dataLabels = NULL, ...)
}
\arguments{
\item{size}{Numeric. Manual size of the radialBars instead of calculating automatically from default height / width.}

\item{inverseOrder}{Logical. Whether to make the first value of series innermost or outermost.}

\item{startAngle}{Numeric. Angle from which the radialBars should start.}

\item{endAngle}{Numeric. Angle to which the radialBars should end. The sum of the startAngle and endAngle should not exceed 360.}

\item{offsetX}{Numeric. Sets the left offset for radialBars.}

\item{offsetY}{Numeric. Sets the top offset for radialBars.}

\item{hollow}{List.}

\item{track}{List.}

\item{dataLabels}{List.}

\item{...}{Additional parameters.}
}
\description{
Use these options in \code{\link{ax_plotOptions}}.
}
\note{
See \url{https://apexcharts.com/docs/options/plotoptions/radialbar/}.
}
