% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input.R
\name{run_demo_input}
\alias{run_demo_input}
\title{Run Shiny input events examples}
\usage{
run_demo_input(example = c("click", "zoom", "selection"))
}
\arguments{
\item{example}{Name of the example.}
}
\description{
Run Shiny input events examples
}
\examples{
if (interactive()) {

  run_demo_input("click")
  run_demo_input("zoom")
  run_demo_input("selection")

}
}
