% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input.R
\name{run_demo_sparkbox}
\alias{run_demo_sparkbox}
\title{Run Shiny spark boxes example}
\usage{
run_demo_sparkbox()
}
\description{
Run Shiny spark boxes example
}
\examples{
if (interactive()) {

  run_sparkbox_demo()

}
}
