% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input.R
\name{run_demo_sync}
\alias{run_demo_sync}
\title{Run Shiny synchronization example}
\usage{
run_demo_sync()
}
\description{
Run Shiny synchronization example
}
\examples{
if (interactive()) {

  run_demo_sync()

}
}
