% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input.R
\name{run_input_demo}
\alias{run_input_demo}
\title{Run Shiny input events examples}
\usage{
run_input_demo(example = c("click", "zoom", "selection"))
}
\arguments{
\item{example}{Name of the example.}
}
\description{
Run Shiny input events examples
}
\examples{
if (interactive()) {

  run_input_demo("click")
  run_input_demo("zoom")
  run_input_demo("selection")

}
}
