% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input.R
\name{run_sync_demo}
\alias{run_sync_demo}
\title{Run Shiny synchronization example}
\usage{
run_sync_demo()
}
\description{
Run Shiny synchronization example
}
\examples{
if (interactive()) {

  run_sync_demo()

}
}
