% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-options.R
\name{scrollerOpts}
\alias{scrollerOpts}
\title{Options for scroller}
\usage{
scrollerOpts(enabled = NULL, height = NULL, track = NULL,
  thumb = NULL, scrollButtons = NULL, padding = NULL,
  offsetX = NULL, offsetY = NULL, ...)
}
\arguments{
\item{enabled}{}

\item{height}{}

\item{track}{}

\item{thumb}{}

\item{scrollButtons}{}

\item{padding}{}

\item{offsetX}{}

\item{offsetY}{}

\item{...}{}
}
\description{
Options for scroller
}
