% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-options.R
\name{selectionOpts}
\alias{selectionOpts}
\title{Selection options}
\usage{
selectionOpts(enabled = NULL, type = NULL, selectedPoints = NULL,
  fill = NULL, stroke = NULL, xaxis = NULL, yaxis = NULL, ...)
}
\arguments{
\item{enabled}{}

\item{type}{}

\item{selectedPoints}{}

\item{fill}{}

\item{stroke}{}

\item{xaxis}{}

\item{yaxis}{}

\item{...}{}
}
\description{
Selection options
}
