% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input.R
\name{set_input_category}
\alias{set_input_category}
\title{Retrieve category (x-axis) in Shiny}
\usage{
set_input_category(ax, inputId)
}
\arguments{
\item{ax}{An \code{apexcharts} \code{htmlwidget} object.}

\item{inputId}{The id that will be used server-side for retrieveng category.}
}
\value{
An \code{apexcharts} \code{htmlwidget} object.
}
\description{
Retrieve category (x-axis) in Shiny
}
