% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input.R
\name{set_input_zoom}
\alias{set_input_zoom}
\title{Retrieve zoom information in Shiny}
\usage{
set_input_zoom(ax, inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{ax}{An \code{apexcharts} \code{htmlwidget} object.}

\item{inputId}{The id that will be used server-side for retrieving zoom.}

\item{session}{The Shiny session.}
}
\value{
An \code{apexcharts} \code{htmlwidget} object.
}
\description{
Retrieve zoom information in Shiny
}
\note{
If x-axis is of type datetime, value retrieved is of class \code{POSIXct}.
}
\examples{
if (interactive()) {

  run_input_demo("zoom")

}
}
