% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperatures}
\alias{temperatures}
\title{Temperature data}
\format{
A data frame with 365 observations and 6 variables.
}
\source{
Enedis (\url{https://data.enedis.fr/explore/dataset/donnees-de-temperature-et-de-pseudo-rayonnement/})
}
\usage{
temperatures
}
\description{
The dataset contains data about temperatures in France between 2018 and 2022.
}
\keyword{datasets}
