% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apexcharter.R
\name{apexchart}
\alias{apexchart}
\title{Create a apexcharts.js widget}
\usage{
apexchart(ax_opts = list(), data = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{ax_opts}{A \code{list} in JSON format with chart parameters}

\item{data}{A \code{data.frame}.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Create a apexcharts.js widget
}
\examples{

library(apexcharter)

apexchart(ax_opts = list(
  chart = list(type = "bar"),
  series = list(list(
    name = "Example",
    data = sample(1:100, 5)
  )),
  xaxis = list(categories = LETTERS[1:5])
))
}
