% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apexcharter-package.R
\docType{package}
\name{apexcharter-package}
\alias{apexcharter-package}
\title{An \code{htmlwidget} interface to the
ApexCharts javascript chart library}
\description{
This package allow you to use ApexCharts.js (\url{https://apexcharts.com/}),
to create interactive and modern SVG charts.
}
\author{
Victor Perrier (@dreamRs_fr)
}
