% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_annotations}
\alias{ax_annotations}
\title{Annotations properties}
\usage{
ax_annotations(ax, position = NULL, yaxis = NULL, xaxis = NULL,
  points = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{position}{Whether to put the annotations behind the charts or in front of it. Available Options: \code{"front"} or \code{"back"}.}

\item{yaxis}{List of lists.}

\item{xaxis}{List of lists.}

\item{points}{List of lists.}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Annotations properties
}
\note{
See \url{https://apexcharts.com/docs/options/annotations/}.
}
