% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_fill}
\alias{ax_fill}
\title{Fill property}
\usage{
ax_fill(ax, type = NULL, colors = NULL, opacity = NULL, gradient = NULL,
  image = NULL, pattern = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{type}{}

\item{colors}{}

\item{opacity}{}

\item{gradient}{}

\item{image}{}

\item{pattern}{}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Fill property
}
