% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex-utils.R
\name{ax_plotOptions}
\alias{ax_plotOptions}
\title{Specific options for chart}
\usage{
ax_plotOptions(ax, bar = NULL, heatmap = NULL, radialBar = NULL,
  pie = NULL, ...)
}
\arguments{
\item{ax}{A \code{apexcharts} \code{htmlwidget} object.}

\item{bar}{See \code{\link{barOpts}}.}

\item{heatmap}{See \code{\link{heatmapOpts}}.}

\item{radialBar}{See \code{\link{radialBarOpts}}.}

\item{pie}{See \code{\link{pieOpts}}.}

\item{...}{Additional parameters.}
}
\value{
A \code{apexcharts} \code{htmlwidget} object.
}
\description{
Specific options for chart
}
